chore: make clippy happy

+++ singlestage/src/components/carousel/carousel.rs
@@ -143,7 +143,7 @@ pub fn Carousel(
             let width = ul_ref.scroll_width();
             let scroll_step = width / num_items;
 
-            ul_ref.set_scroll_left(scroll_step * (cur - 1) as i32);
+            ul_ref.set_scroll_left(scroll_step * (cur - 1));
         }
     });
